Extract Ktor client factory into core data
