+++ app/src/main/java/timur/gilfanov/messenger/di/NetworkModule.kt
@@ -6,17 +6,6 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.HttpClientEngine
-import io.ktor.client.engine.cio.CIO
-import io.ktor.client.plugins.HttpRequestRetry
-import io.ktor.client.plugins.HttpTimeout
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.plugins.defaultRequest
-import io.ktor.client.plugins.logging.LogLevel
-import io.ktor.client.plugins.logging.Logger
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
-import io.ktor.serialization.kotlinx.json.json
 import javax.inject.Named
 import javax.inject.Qualifier
 import javax.inject.Singleton
@@ -24,6 +13,7 @@ import kotlinx.serialization.json.Json
 import timur.gilfanov.messenger.BuildConfig
 import timur.gilfanov.messenger.auth.AuthInterceptor
 import timur.gilfanov.messenger.auth.di.UnauthenticatedHttpClient
+import timur.gilfanov.messenger.data.remote.network.KtorHttpClientFactory
 
 /**
  * Qualifier for the base URL configuration
@@ -47,11 +37,6 @@ annotation class MockHttpClient
 @InstallIn(SingletonComponent::class)
 object NetworkModule {
 
-    private const val MAX_CONNECTIONS_COUNT = 100
-    private const val REQUEST_TIMEOUT_MILLIS = 30_000L
-    private const val CONNECT_TIMEOUT_MILLIS = 10_000L
-    private const val SOCKET_TIMEOUT_MILLIS = 30_000L
-
     /**
      * Provides the base URL for API endpoints.
      * Can be configured through BuildConfig for different environments.
@@ -66,13 +51,7 @@ object NetworkModule {
      */
     @Provides
     @Singleton
-    fun provideJson(): Json = Json {
-        prettyPrint = true
-        isLenient = true
-        ignoreUnknownKeys = true
-        encodeDefaults = true
-        coerceInputValues = true
-    }
+    fun provideJson(): Json = KtorHttpClientFactory.createJson()
 
     /**
      * Provides the HTTP client engine.
@@ -80,14 +59,7 @@ object NetworkModule {
      */
     @Provides
     @Singleton
-    fun provideHttpClientEngine(): HttpClientEngine = CIO.create {
-        // Connection configuration
-        maxConnectionsCount = MAX_CONNECTIONS_COUNT
-        pipelining = true
-
-        // Request timeout configuration
-        requestTimeout = REQUEST_TIMEOUT_MILLIS
-    }
+    fun provideHttpClientEngine(): HttpClientEngine = KtorHttpClientFactory.createEngine()
 
     /**
      * Provides the main HTTP client for production use.
@@ -102,7 +74,9 @@ object NetworkModule {
         @BaseUrl baseUrl: String,
         logger: timur.gilfanov.messenger.util.Logger,
         authInterceptor: AuthInterceptor,
-    ): HttpClient = httpClient(engine, json, baseUrl, logger).also { authInterceptor.install(it) }
+    ): HttpClient = createHttpClient(engine, json, baseUrl, logger).also {
+        authInterceptor.install(it)
+    }
 
     /**
      * Provides an HTTP client without authentication interceptor.
@@ -118,41 +92,20 @@ object NetworkModule {
         json: Json,
         @BaseUrl baseUrl: String,
         logger: timur.gilfanov.messenger.util.Logger,
-    ): HttpClient = httpClient(engine, json, baseUrl, logger)
+    ): HttpClient = createHttpClient(engine, json, baseUrl, logger)
 
-    private fun httpClient(
+    private fun createHttpClient(
         engine: HttpClientEngine,
         json: Json,
         baseUrl: String,
         logger: timur.gilfanov.messenger.util.Logger,
-    ): HttpClient = HttpClient(engine) {
-        defaultRequest {
-            this.url(baseUrl)
-            contentType(ContentType.Application.Json)
-        }
-        this.install(HttpRequestRetry) {
-            retryOnServerErrors(maxRetries = 3)
-            retryOnException(maxRetries = 3, retryOnTimeout = true)
-            exponentialDelay()
-        }
-        this.install(HttpTimeout) {
-            requestTimeoutMillis = REQUEST_TIMEOUT_MILLIS
-            connectTimeoutMillis = CONNECT_TIMEOUT_MILLIS
-            socketTimeoutMillis = SOCKET_TIMEOUT_MILLIS
-        }
-        this.install(ContentNegotiation) {
-            json(json)
-        }
-        if (BuildConfig.DEBUG) {
-            this.install(Logging) {
-                this.logger = object : Logger {
-                    override fun log(message: String) {
-                        logger.d("NetworkModule", message)
-                    }
-                }
-                level = LogLevel.ALL
-            }
-        }
+    ): HttpClient = KtorHttpClientFactory.create(
+        engine = engine,
+        json = json,
+        baseUrl = baseUrl,
+        enableLogging = BuildConfig.DEBUG,
+    ) { message ->
+        logger.d("NetworkModule", message)
     }
 
     /**
@@ -161,10 +114,6 @@ object NetworkModule {
      */
     @Provides
     @Named("mock")
-    fun provideMockHttpClient(json: Json): HttpClient = HttpClient {
-        // Basic configuration for mock client
-        install(ContentNegotiation) {
-            json(json)
-        }
-    }
+    fun provideMockHttpClient(json: Json): HttpClient =
+        KtorHttpClientFactory.createContentNegotiationClient(json)
 }
